+++ nexus-ingestion-api/src/main/java/org/techbd/NexusIngestionApiApplication.java
@@ -1,4 +1,4 @@
-package org.techbd.nexusingestionapi;
+package org.techbd;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
