+++ nexus-ingestion-api/src/test/java/org/techbd/controller/DataIngestionControllerTest.java
@@ -1,4 +1,4 @@
-package org.techbd.nexusingestionapi.controller;
+package org.techbd.controller;
 
 import static org.junit.jupiter.api.Assertions.*;
 import static org.mockito.Mockito.*;
@@ -12,7 +12,7 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-import org.techbd.nexusingestionapi.service.AwsService;
+import org.techbd.service.AwsService;
 
 import software.amazon.awssdk.services.sqs.SqsClient;
