--- nexus-ingestion-api/src/test/java/org/techbd/nexusingestionapi/NexusIngestionApiApplicationTests.java
@@ -1,13 +1,0 @@
-package org.techbd.nexusingestionapi;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class NexusIngestionApiApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
